Дорабатывает вывод in/to для точек и формы
